refactor(coil-extension): extract backgroundMain
